disable custom certificate path in fips mode
