+++ Makefile
@@ -36,9 +36,11 @@ check-vet:
 docker-test:
 	scripts/docker-compose-testing run -T --rm go-agent-tests make test
 
-.PHONY: test
+.PHONY: test-fips test
+test-fips: ARGS=-tags=requirefips
+test-fips: test
 test:
-	@for dir in $(shell scripts/moduledirs.sh); do (cd $$dir && go test -race -v -timeout=$(TEST_TIMEOUT) ./...) || exit $$?; done
+	@for dir in $(shell scripts/moduledirs.sh); do (cd $$dir && go test -race -v -timeout=$(TEST_TIMEOUT) $(ARGS) ./...) || exit $$?; done
 
 .PHONY: coverage
 coverage:
